Move adjacency constructors to module
